Rename list package to cashflow for clarity
